Isolate PropertyItem on a separate component
