+++ src/modules/CRUD/components/Details/PropertyItem/index.js
@@ -1,0 +1,8 @@
+import { withNamespaces } from 'react-i18next';
+import compose from '../../../../../utils/compose';
+
+import PropertyItem from './PropertyItem';
+
+export default compose(
+  withNamespaces(),
+)(PropertyItem);
